+++ app/src/main/java/com/johnsimon/payback/util/Resource.java
@@ -321,14 +321,14 @@ public class Resource {
         public int compare(Debt a, Debt b) {
             if (a.isPaidBack()) {
                 if(b.isPaidBack()) {
-                    return Math.round(b.getAbsoluteAmount() - a.getAbsoluteAmount());
+                    return (int) Math.round(b.getAbsoluteAmount() - a.getAbsoluteAmount());
 				} else {
 					return 1;
 				}
 			} else if(b.isPaidBack()) {
 				return -1;
 			} else {
-				return Math.round(b.getAbsoluteAmount() - a.getAbsoluteAmount());
+				return (int) Math.round(b.getAbsoluteAmount() - a.getAbsoluteAmount());
 			}
         }
     }
